requirements.txt
numpy>=1.26.0
pandas>=2.0.0
pyarrow>=14.0.0
streamlit>=1.28.0
scikit-learn>=1.3.0
openai
psycopg2-binary
python-dotenv
fastapi
uvicorn
apscheduler
requests
toml
PyJWT>=2.8.0
